// backend/Types/enums.ts
export enum VAssetType {
	Crypto = "Crypto",
	Equity = "Equity",
	Commodity = "Commodity",
	FixedIncome = "Fixed Income",
	RealEstate = "Real Estate",
	Cash = "Cash",
	Other = "Other",
}

export enum VTransactionType {
	Buy = "Buy",
	Sell = "Sell",
}

export enum VConnectionType {
	EthereumWallet = 'EthereumWallet',
	AvalancheWallet = 'AvalancheWallet',
	AvalaunchIDOContract ='AvalaunchIDOContract',
	BinanceAccount = 'BinanceAccount',
	Kucoin = 'Kucoin',
	Okx = 'Okx',
	Hitbtc = 'Hitbtc',
	Cryptocom = 'Cryptocom',
	Coinbase = 'Coinbase',
	MEXC = 'MEXC',
	FTX = 'FTX',
	BitFinex = 'BitFinex',
	Bittrex = 'Bittrex',
	Huobi = 'Huobi',
	Kraken = 'Kraken',
	Poloniex = 'Poloniex',
	Gate = 'Gate',
	Ascendex = 'Ascendex'
}